loan_rejector/requirements.txt
easyocr
numpy
Pillow
openai
sentence-transformers
python-dotenv
